requirements.20211205.txt
appnope==0.1.2
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bs4==0.0.1
cached-property==1.5.2
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
commonmark==0.9.1
decorator==5.1.0
dictdiffer==0.9.0
idna==3.3
IMDbPY==2021.4.18
ipython==7.30.1
ipython-genutils==0.2.0
jedi==0.18.1
lxml==4.6.4
matplotlib-inline==0.1.3
notion @ git+https://github.com/jamalex/notion-py.git@017de43af765e232f6ba4403c5fa3513e614c955
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.23
ptyprocess==0.7.0
Pygments==2.10.0
python-slugify==5.0.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-103/six-1.15.0-py2.py3-none-any.whl
soupsieve==2.3.1
SQLAlchemy==1.4.27
text-unidecode==1.3
traitlets==5.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
wcwidth==0.2.5
